refactor(api-sdk): support flexible email identifiers for update
